+++ src/components/modals/GeneralModal.tsx
@@ -32,6 +32,7 @@ export default function GeneralModal({
                     "fixed inset-0 bg-black bg-opacity-30 backdrop-blur-sm flex justify-center items-center"
             }
             style={{zIndex:999}}
+            role={"dialog"}
         >
             <div className="bg-white p-4 rounded dark:bg-gray-900 min-w-[50vw]">
                 <h1 className="font-semibold text-center text-xl text-gray-700 mb-4 dark:text-gray-200">