+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Features
 
 * Multiselect box changed to List of checkboxes in Service forms
+* TableViewComponent support multi selection
+* Import Shop Data dialog added to Shop list
 
 
 ## [1.1.0] (2024-11-04)